+++ projects/ontimize-web-ngx/src/lib/components/grid/o-grid.component.ts
@@ -70,7 +70,7 @@ export const DEFAULT_INPUTS_O_GRID = [
   'insertButtonPosition:insert-button-position',
   // insert-button-floatable [no|yes]: Indicates whether or not to position of the insert button is floating . Default: 'yes'
   'insertButtonFloatable:insert-button-floatable',
-  'quickfilterAppearance:quickfilter-appearance'
+  'quickFilterAppearance:quick-filter-appearance'
 ];
 
 export const DEFAULT_OUTPUTS_O_GRID = [
@@ -128,7 +128,7 @@ export class OGridComponent extends AbstractOServiceComponent<OGridComponentStat
   public insertButtonPosition: 'top' | 'bottom' = 'bottom';
 
   public gutterSize = '1px';
-  protected _quickfilterAppearance: MatFormFieldAppearance = 'outline';
+  protected _quickFilterAppearance: MatFormFieldAppearance = 'outline';
   get cols(): number {
     return this._cols || this._colsDefault;
   }
@@ -532,16 +532,16 @@ export class OGridComponent extends AbstractOServiceComponent<OGridComponentStat
       this.quickFilterComponent.setActiveColumns(parsedArr);
     }
   }
-  get quickfilterAppearance(): MatFormFieldAppearance {
-    return this._quickfilterAppearance;
+  get quickFilterAppearance(): MatFormFieldAppearance {
+    return this._quickFilterAppearance;
   }
 
-  set quickfilterAppearance(value: MatFormFieldAppearance) {
+  set quickFilterAppearance(value: MatFormFieldAppearance) {
     const values = ['legacy', 'standard', 'fill', 'outline'];
     if (values.indexOf(value) === -1) {
-      console.warn('The quickfilterAppearance attribute is undefined');
+      console.warn('The quickFilterAppearance attribute is undefined');
       value = 'outline';
     }
-    this._quickfilterAppearance = value;
+    this._quickFilterAppearance = value;
   }
 }